added seperate styles.js for each component
