+++ client/src/components/Footer/Footer.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { StyledFooter } from './styles';
+
+function Footer() {
+  return (
+    <StyledFooter>
+      <p>All Rights Reserved. Nicholas Chumney.</p>
+    </StyledFooter>
+  );
+}
+
+export default Footer;
