+++ client/src/components/Footer/styles.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+export const StyledFooter = styled.footer`
+  width: 100%;
+  height: 100%;
+  padding: 0 2rem;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background: #203040;
+  color: #ffffff;
+`;
