+++ client/src/pages/Landing/Landing.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { PageContent } from './styles';
+
+function Landing() {
+  return (
+    <PageContent>
+      <h1>Landing Page</h1>
+    </PageContent>
+  );
+}
+
+export default Landing;
