+++ client/src/pages/NotFound/styles.js
@@ -1,0 +1,8 @@
+import styled from 'styled-components';
+
+export const PageContent = styled.div`
+  width: 100%;
+  height: 100%;
+  min-height: 768px;
+  padding: 2rem;
+`;
