+++ client/src/pages/Register/Register.js
@@ -1,68 +1,20 @@
 import React, { useState } from 'react';
 import { Link } from 'react-router-dom';
-import styled from 'styled-components';
 import axios from 'axios';
 import jwt_decode from 'jwt-decode';
+import {
+  PageContent,
+  StyledForm,
+  StyledFormGroup,
+  StyledFormHeader,
+  StyledButton,
+} from './styles';
 
-const StyledLogin = styled.div`
-  width: 100%;
-  height: 100%;
-  padding-top: 4rem;
-`;
-
-const StyledForm = styled.form`
-  width: 50%;
-  max-width: 600px;
-  margin: 0 auto;
-  padding: 2rem;
-  border: 1px solid #000;
-  & p {
-    text-align: center;
-  }
-`;
-
-const StyledFormHeader = styled.div`
-  margin-bottom: 1rem;
-  text-align: center;
-  & h2 {
-    font-size: 1.5rem;
-  }
-  & p {
-    margin: 1rem 0;
-    background: red;
-    color: #000;
-  }
-`;
-
-const StyledFormGroup = styled.div`
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 1rem;
-  & label {
-    margin-bottom: 0.3rem;
-    font-size: 1rem;
-  }
-  & input {
-    padding: 0.8rem;
-  }
-  & small {
-    font-size: 0.8rem;
-    color: red;
-  }
-`;
-
-const StyledButton = styled.button`
-  width: 100%;
-  margin: 1rem auto;
-  padding: 0.8rem 1rem;
-  font-size: 1rem;
-  cursor: pointer;
-`;
-
-function Login() {
+function Register() {
+  const [name, setName] = useState('');
   const [email, setEmail] = useState('');
   const [password, setPassword] = useState('');
+  const [password2, setPassword2] = useState('');
   const [isLoading, setLoading] = useState(false);
   const [errors, setErrors] = useState({});
 
@@ -71,9 +23,11 @@ function Login() {
     setLoading(true);
 
     try {
-      const res = await axios.post('http://localhost:8081/api/users/login', {
+      const res = await axios.post('http://localhost:8081/api/users/register', {
+        name,
         email,
         password,
+        password2,
       });
 
       const user = jwt_decode(res.data.token);
@@ -96,12 +50,25 @@ function Login() {
   };
 
   return (
-    <StyledLogin>
+    <PageContent>
       <StyledForm onSubmit={handleSubmit}>
         <StyledFormHeader>
-          <h2>Sign in to your account</h2>
+          <h2>Start an account</h2>
           {errors.message && <p>{errors.message}</p>}
         </StyledFormHeader>
+        <StyledFormGroup>
+          <label htmlFor="name">Name</label>
+          <input
+            required
+            type="text"
+            id="name"
+            name="name"
+            placeholder="Enter your name"
+            value={name}
+            onChange={(e) => setName(e.target.value)}
+          />
+          {errors.name && <small>{errors.name}</small>}
+        </StyledFormGroup>
         <StyledFormGroup>
           <label htmlFor="email">Email</label>
           <input
@@ -128,16 +95,29 @@ function Login() {
           />
           {errors.password && <small>{errors.password}</small>}
         </StyledFormGroup>
+        <StyledFormGroup>
+          <label htmlFor="password2">Confirm password</label>
+          <input
+            required
+            type="password"
+            id="password2"
+            name="password2"
+            placeholder="Confirm new password"
+            value={password2}
+            onChange={(e) => setPassword2(e.target.value)}
+          />
+          {errors.password2 && <small>{errors.password2}</small>}
+        </StyledFormGroup>
 
         <StyledButton type="submit" disabled={isLoading}>
-          Login
+          Register
         </StyledButton>
         <p>
-          Don't have an account? <Link to="/register">Sign Up</Link>
+          Already have an account? <Link to="/login">Sign In</Link>
         </p>
       </StyledForm>
-    </StyledLogin>
+    </PageContent>
   );
 }
 
-export default Login;
+export default Register;
